Template: Don't duplicate template locating
References:
- https://developer.wordpress.org/reference/functions/locate_template/
- https://developer.wordpress.org/reference/functions/get_template_part/

Previously, this code would call locate_template(), (which would check for a file in a child theme, or parent theme, or in a wp-includes/theme-compat/ directory, and if it was found, then it would return true, but wouldn't load anything.

It would then call get_template_part(), and along with a couple of filters, would try to establish the filename (in this case, a predefined file name from the slug with `.php` appended, same as the arg to `locate_template()`), then it would call `locate_template()` (which checks the theme locations again), but this time it would load it.

If the file was not originally found in the first `locate_template()`, it would return false, and the fallback to including the file from the plugin would happen from the `else`.

The change in this commit simplifies things. We now start off with the `get_template_part()` call (which, in a future commit, could have filterable default args), which determines the file name to look for from the provided args, and then calls `locate_template()` once to see if it exists (and load if so), the file. If it doesn't exist, `get_template_part()` returns false, so our default template within the plugin can be included.