Template: Allow for custom template name flexibility
Until now, the instructions have been that the custom template MUST be called `template-maintenance-mode.php`, and it MUST be located in the root of the theme (child, parent, or in a `wp-includes/theme-compat/` directory).

By making use of the `get_template_part()` args, and providing a filter, we can add flexibility to theme developers for where they want a custom template to live.

For instance, if they want it to live at `wp-content/themes/my-theme/plugin-templates/maintenance-mode-alt.php`, then they could add some code like:

```
add_filter(
    'vip_maintenance_mode_template_args',
    function( $args ) {
        $args['slug'] = 'plugin-templates/maintenance-mode';
        $args['name'] = 'alt';

        return $args;
    }
);
```

This also allows third key to be used to pass in custom arguments to the template.